refactor: Adjust internship offer registration to use authentication.
