+++ src/main/java/br/edu/ifpb/pweb2/estagion/repositories/CompanyRepository.java
@@ -10,4 +10,5 @@ import java.util.Optional;
 public interface CompanyRepository extends JpaRepository<Company, Integer> {
     Optional<Company> findByEmail(String email);
     Optional<Company> findByCnpj(String cnpj);
+    Optional<Company> findByUsername(String username);
 }
