+++ src/main/java/br/edu/ifpb/pweb2/estagion/service/CompanyService.java
@@ -2,6 +2,7 @@ package br.edu.ifpb.pweb2.estagion.service;
 
 import br.edu.ifpb.pweb2.estagion.model.Authority;
 import br.edu.ifpb.pweb2.estagion.model.Company;
+import br.edu.ifpb.pweb2.estagion.model.InternshipOffer;
 import br.edu.ifpb.pweb2.estagion.repositories.AuthorityRepository;
 import br.edu.ifpb.pweb2.estagion.repositories.CompanyRepository;
 import jakarta.transaction.Transactional;
@@ -23,17 +24,29 @@ public class CompanyService {
     @Autowired
     private AuthorityRepository authorityRepository;
 
+    @Transactional
     public List<Company> findAll() {
         return repository.findAll();
     }
 
+    @Transactional
     public Company findById(int id) {
         return repository.findById(id).orElse(null);
     }
 
     @Transactional
+    public Company findByUsername(String username) {
+        return repository.findByUsername(username).orElse(null);
+    }
+
+    @Transactional
     public Company findByCnpj(String cnpj) {
         return repository.findByCnpj(cnpj).orElse(null);
+    }
+
+    public void saveInternshipOffer(Company company, InternshipOffer offer) {
+        company.getInternshipOffers().add(offer);
+        repository.save(company);
     }
 
     @Transactional
